GAME WORKS
Pollet is dumb.  If skipped players don’t enter history, there’s no way
to tell from a state who’s turn it is.  So state is incomplete
information, because we rely on the “who played this card” to tell
who’s turn it is, but when you’re skipped that stat becomes useless.
So I implemented it where when you’re skipped you just play nothing and
draw nothing, but that move is added to history.